requirements.txt
fastapi==0.99.0
pydantic==1.10.9
groq==0.4.0
python-dotenv==1.0.0
uvicorn==0.23.2
